refactor: 重新設計 gun system 的資料夾位置
